src/requirements.txt
langchain 
langchain_community 
langchain-openai 
langgraph 
ipython
python-dotenv
fastapi
uvicorn
langfuse
gradio